Blood System Management

+++ src/main/java/com/digitalkrapht/bloodbank/bloodbank/organization/repositroy/BloodCollectionRepository.java
@@ -1,14 +1,14 @@
 package com.digitalkrapht.bloodbank.bloodbank.organization.repositroy;
 
 import com.digitalkrapht.bloodbank.bloodbank.organization.models.BloodCollection;
-import com.digitalkrapht.bloodbank.bloodbank.organization.models.BloodGroup;
+import com.digitalkrapht.bloodbank.bloodbank.organization.models.BloodCollectionLog;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface BloodCollectionRepository  extends JpaRepository<BloodCollection,Integer> {
+public interface BloodCollectionRepository  extends JpaRepository<BloodCollection,Long> {
     List<BloodCollection> findByEnabled(boolean status);
     Page<BloodCollection> findByEnabled(boolean status, Pageable pageable);
     Long countByEnabled(boolean status);
