+++ src/main/java/com/digitalkrapht/bloodbank/bloodbank/organization/request/AddStockRequest.java
@@ -1,13 +1,8 @@
 package com.digitalkrapht.bloodbank.bloodbank.organization.request;
 
-import com.digitalkrapht.bloodbank.bloodbank.organization.models.BloodGroup;
-import com.digitalkrapht.bloodbank.bloodbank.organization.models.StockAdjustmentType;
-import com.digitalkrapht.bloodbank.bloodbank.users.models.UserBackOfficeAgent;
-import com.digitalkrapht.bloodbank.bloodbank.users.models.UserDonor;
+import com.digitalkrapht.bloodbank.bloodbank.organization.models.enums.StockAdjustmentType;
 import lombok.Data;
 
-import javax.persistence.JoinColumn;
-import javax.persistence.OneToOne;
 import javax.validation.constraints.NotNull;
 
 @Data
@@ -21,8 +16,8 @@ public class AddStockRequest {
     @NotNull
     private String backAgentId;
     @NotNull
-    private int bloodGroupId;
+    private long bloodGroupId;
 
-    ;
+
 
 }
