+++ src/main/java/com/digitalkrapht/bloodbank/bloodbank/organization/request/UpdateStockDetailsRequest.java
@@ -1,13 +1,13 @@
 package com.digitalkrapht.bloodbank.bloodbank.organization.request;
 
-import com.digitalkrapht.bloodbank.bloodbank.organization.models.StockAdjustmentType;
+import com.digitalkrapht.bloodbank.bloodbank.organization.models.enums.StockAdjustmentType;
 import lombok.Data;
 
 import javax.validation.constraints.NotNull;
 
 @Data
 public class UpdateStockDetailsRequest {
-    private int id;
+    private long id;
     private String quantity;
     private String unit;
     private StockAdjustmentType stockAdjustmentType;
@@ -17,5 +17,5 @@ public class UpdateStockDetailsRequest {
     @NotNull
     private String backAgentId;
     @NotNull
-    private int bloodGroupId;
+    private long bloodGroupId;
 }
